refactor: move FileTask from pipeline to crate root so shared type lives at the API surface
